+++ package.json
@@ -36,6 +36,7 @@
     "react-native-google-signin": "^2.1.1",
     "react-native-material-ui": "^1.30.1",
     "react-native-modal": "^11.5.6",
+    "react-native-multiple-select": "^0.5.5",
     "react-native-reanimated": "~1.4.0",
     "react-native-safe-area-context": "0.6.0",
     "react-native-screens": "2.0.0-alpha.12",
